api_yamdb/requirements.txt
asgiref==3.2.10

Django==2.2.16

djangorestframework==3.12.4
django-debug-toolbar
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-filter==2.4.0

gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1

PyJWT==2.1.0

pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3

requests==2.26.0

sqlparse==0.3.1 
